Created the TicketModifier.php files
This is a concept file, very early in the creation process. If a uid and tid are supplied through GET, it will display a ticket modifying form for that ticket. If not, it will display a ticket creation form for this user. Results are sent to 'Dashboard.php' and 'Suggestions.php' respectively.